requirements.txt
# Core dependencies
torch>=2.0.0
transformers>=4.30.0
datasets>=2.14.0
accelerate>=0.20.0

# Configuration and utilities
pyyaml>=6.0
numpy>=1.24.0
tqdm>=4.65.0

# Logging and monitoring
tensorboard>=2.13.0
wandb>=0.15.0

# Data processing
jsonlines>=3.1.0

